edit 2 add 100-largets prime number

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -11,7 +11,7 @@ void prime_factor(void)
 {
 	int i;
 	long int maxPrime = 1;
-	int num = 612852475143;
+	long num = 612852475143;
 
 	while (num % 2 == 0)
 	{
